[REFACTOR] add notificationHandlers
